+++ CHANGELOG.md
@@ -10,6 +10,13 @@ FaasJS use [Semantic Versioning](https://semver.org/).
 - [Break] Remove `@faasjs/mongo`, `@faasjs/vue-plugin`.
 - [Break] Replace jest with vitest.
 
+- `@faasjs/server`
+  - [Feature] Add logger to server's handler.
+  - [Feature] Stop logger transport when server close.
+
+- `@faasjs/logger`
+  - [Feature] Add reset method to logger transport.
+
 [`v4.7.2 (2024-12-25)`](https://github.com/faasjs/faasjs/compare/v4.7.0...v4.7.2)
 
 - `@faasjs/logger`